refactor(#725): move the row-less release synthesis cluster out of the orchestrator
Pure code motion, no behavior change (PR 3 of the #722 orchestrator decomposition). The row-less / non-library-release synthesis cluster (#628/#631/#632/#660/#663 lineage) moves verbatim from lookup/orchestrator.py to the new lookup/rowless.py: the LML#663 packed-title credit recovery (_PACKED_TITLE_SEPARATOR, _own_release_credit), the LML#631 SONG_AS_ARTIST row-less pick (_select_rowless_artist_release), the synthetic id=0 chokepoint (ROWLESS_LIBRARY_ID, NONLIBRARY_RELEASE_SURFACED_STAT_KEY, _make_rowless_item), the #628 A1 carry-through kernel (_resolve_nonlibrary_release) with its #632 cache re-hydrate (_rehydrate_resolved_release), and the LML#660 per-track credit recovery (_MAX_CREDIT_RECOVERY_FETCHES, _recover_track_credit). Two constants sitting inside the block stay behind with their consumers: SONG_AS_TRACK_CONFIDENCE (read only by _match_track_releases_to_library, which moves in PR 4a) and ROWLESS_NO_ALBUM_CONFIDENCE (read only by find_library_albums_with_cached_track and _bind_resolved_release, which move in PR 5). The orchestrator imports only the names its remaining code consumes (ROWLESS_LIBRARY_ID, _make_rowless_item, _recover_track_credit, _resolve_nonlibrary_release, _select_rowless_artist_release); the now-unconsumed entity.release_resolution_cache block, resolve_release_for_track, and get_cache_stats_recorder imports are pruned. No re-exports.
